027 - Add simple camera

+++ Engine/Bey3D/include/Plane.h
@@ -4,6 +4,7 @@
 #include <array>
 #include "IndexedTriangleList.h"
 #include "BeyMath.h"
+#include "DirectXMath.h"
 
 class Plane
 {
@@ -46,9 +47,10 @@ public:
 			{
 				return static_cast<unsigned short>(y * nVertices_x + x);
 			};
-			for (size_t y = 0; y < divisions_y; y++)
+
+			for (size_t y = 0; y < size_t(divisions_y); y++)
 			{
-				for (size_t x = 0; x < divisions_x; x++)
+				for (size_t x = 0; x < size_t(divisions_x); x++)
 				{
 					const std::array<unsigned short, 4> indexArray =
 					{ vxy2i(x,y),vxy2i(x + 1,y),vxy2i(x,y + 1),vxy2i(x + 1,y + 1) };
